Added and verified support for prismatic joint

+++ simulation/sample_shapes.py
@@ -2,6 +2,7 @@
 Code to sample points from various 3D rigid body, when one samples only points from a plane or line, the resulting motion estimate is degenerate
 '''
 import numpy as np
+import pdb
 
 '''
 Given parameters corresponding to a particular shape and a shape type, this code will sample some points on that
@@ -15,13 +16,13 @@ def sample_points(shape_pars,shape_type):
         # takes 3 shape parameters as argument: Length of major axis along x,y and z
         samples = np.matrix([[0,0,0],[shape_pars[0],0,0],[shape_pars[0]/2.0,0,shape_pars[2]/2.0],
             [shape_pars[0]/2.0,0,-shape_pars[2]/2.0],[shape_pars[0]/2.0,shape_pars[1]/2.0,0],
-            [shape_pars[0]/2.0,-shape_pars[1]/2.0,0]])
+            [shape_pars[0]/2.0,-shape_pars[1]/2.0,0]],dtype='float')
     elif shape_type=='cuboid':
         assert (shape_pars.shape[0] >= 3),"Need atleast 3 parameters for length, width, breadth"
         # Sampling 8 extreme points on the cuboid surface
         samples = np.matrix([[0,0,0],[shape_pars[0],0,0],[shape_pars[0],shape_pars[1],0],
             [0,shape_pars[1],0],[0,0,shape_pars[2]],[shape_pars[0],0,shape_pars[2]],
-            [shape_pars[0],shape_pars[1],shape_pars[2]],[0,shape_pars[1],shape_pars[2]]])
+            [shape_pars[0],shape_pars[1],shape_pars[2]],[0,shape_pars[1],shape_pars[2]]],dtype='float')
     elif shape_type=='cylinder':
         assert (shape_pars.shape[0] >= 2), "Need atleast two parameters, circle radius and length"
         # Sampling 4 points from one circle and 4 points from another circle end
@@ -30,7 +31,7 @@ def sample_points(shape_pars,shape_type):
         # Need two shape parameters : shape_pars[0] corresponds to circle radius and shape_pars[1] corresponds to length
         samples = np.matrix([[0,-shape_pars[0],0],[0,shape_pars[0],0],[0,0,-shape_pars[0]],
             [0,0,shape_pars[0]],[shape_pars[1],-shape_pars[0],0],[shape_pars[1],shape_pars[0],0],
-            [shape_pars[1],0,-shape_pars[0]],[shape_pars[1],0,shape_pars[0]]])
+            [shape_pars[1],0,-shape_pars[0]],[shape_pars[1],0,shape_pars[0]]],dtype='float')
     else:
         print "Not a valid shape type"
         print "Allowed shape types are cuboid, ellipse and cylinder"
